CTO-108: wire /, /agents, /cost, /attribution to useLivePoll
Each server component still fetches initial data and renders the first paint;
the rendered body moves into a sibling Live.tsx client component that calls
useLivePoll with the same endpoint and re-derives banner/state on each tick.
A LiveIndicator sits alongside the existing StaleBadge in each page header.
URL params (?tag, ?run, ?provider, ?outcome) are baked into the polled
endpoint server-side so deep links keep working.
